+++ source/isaaclab_diffrl/isaaclab_diffrl/tasks/direct/isaaclab_diffrl/__init__.py
@@ -24,3 +24,17 @@ gym.register(
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
     },
 )
+
+
+gym.register(
+    id="Isaac-Cartpole-DiffRL-Newton-v0",
+    entry_point=f"{__name__}.isaaclab_diffrl_newton_env:IsaaclabDiffrlNewtonEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": f"{__name__}.isaaclab_diffrl_newton_env_cfg:IsaaclabDiffrlNewtonEnvCfg",
+        "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
+        "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:PPORunnerCfg",
+        "skrl_amp_cfg_entry_point": f"{agents.__name__}:skrl_amp_cfg.yaml",
+        "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
+    },
+)